Finished Calc
User Story #15
has several decimal places of precision
User Story #13
Two or more operators are entered the the final one is used, with the exception that if the final one is a minus. The minus needs to remain to evaluate as a negative sign.
User Story #11
Multiple decimals are not allowed in the same number.
linked the decimal button to the correct function decimal, instead of numkey.